fix: resolve cargo fmt and clippy warnings for CI
- Remove unnecessary type casts (as f32, as u32)
- Replace manual div_ceil with .div_ceil()
- Collapse nested if-let chains
- Replace needless range loops with enumerate iterators
- Replace &vec![...] with &[...]
- Add #[allow] for intentional too_many_arguments / type_complexity
- Use .is_none_or(), .is_multiple_of()
- Fix homebrew-tap formula path in release workflow

+++ src/quality.rs
@@ -38,8 +38,7 @@ impl SpeedSettings {
         let kmeans_iteration_limit = 1.0 / f64::from(1u32 << (23 - u32::from(effective_speed)));
 
         let feedback_loop_trials = (56 - 9 * i32::from(effective_speed)).max(0) as u16;
-        let max_histogram_entries =
-            ((1 << 17) + (1 << 18) * (10 - u32::from(effective_speed))) as u32;
+        let max_histogram_entries = (1 << 17) + (1 << 18) * (10 - u32::from(effective_speed));
         let input_posterize_bits = if effective_speed >= 8 { 1 } else { 0 };
 
         let mut use_dither_map = if effective_speed <= 6 {
